Moved classes to correct namespace `Ember`
